Move tests from the 'verifier' package to the 'datastructures' one
