+++ src/jvm/test-lincheck-integration/org/jetbrains/lincheck_test/datastructures/linearizable/RendezvousChannelTest.kt
@@ -7,7 +7,7 @@
  * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.jetbrains.lincheck_test.verifier.linearizability
+package org.jetbrains.lincheck_test.datastructures.linearizable
 
 import kotlinx.coroutines.*
 import kotlinx.coroutines.channels.*
