+++ src/jvm/test-lincheck-integration/org/jetbrains/lincheck_test/datastructures/quiescent/LockFreeTaskQueueTest.kt
@@ -8,7 +8,7 @@
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck_test.verifier.quiescent
+package org.jetbrains.lincheck_test.datastructures.quiescent
 
 import org.jetbrains.kotlinx.lincheck_test.AbstractLincheckTest
 import org.jetbrains.lincheck.datastructures.IntGen
