+++ src/jvm/test-lincheck-integration/org/jetbrains/lincheck_test/datastructures/serializable/SerializableQueueTest.kt
@@ -1,14 +1,14 @@
 /*
  * Lincheck
  *
- * Copyright (C) 2019 - 2023 JetBrains s.r.o.
+ * Copyright (C) 2019 - 2025 JetBrains s.r.o.
  *
  * This Source Code Form is subject to the terms of the
  * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
  * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 
-package org.jetbrains.lincheck_test.verifier.serializability
+package org.jetbrains.lincheck_test.datastructures.serializable
 
 import org.jetbrains.lincheck.datastructures.Operation
 import org.jetbrains.lincheck.datastructures.ModelCheckingOptions
